+++ src/components/viewmodel.ts
@@ -1,0 +1,40 @@
+import { useEffect, useState } from 'react';
+import axios from 'axios';
+
+const useProductViewModel = () => {
+  const [products, setProducts] = useState<any[]>([]);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
+  const [searchTerm, setSearchTerm] = useState<string>('');
+
+  useEffect(() => {
+    const fetchProducts = async () => {
+      try {
+        const response = await axios.get('http://localhost:5000/api/products');
+        setProducts(response.data);
+        setLoading(false);
+      } catch (err) {
+        console.error('Error fetching products:', err);
+        setError('Error al cargar los productos');
+        setLoading(false);
+      }
+    };
+
+    fetchProducts();
+  }, []);
+
+  const filteredProducts = products.filter(product =>
+    product.title.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+
+  return {
+    products,
+    loading,
+    error,
+    searchTerm,
+    setSearchTerm,
+    filteredProducts,
+  };
+};
+
+export default useProductViewModel;
